some changes
Add migrations, add validator service, change db

+++ myteam.holiday.WebApi/Controllers/UserController.cs
@@ -1,6 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Cryptography.KeyDerivation;
+using Microsoft.AspNetCore.Mvc;
+using myteam.holiday.WebApi.Services;
 using myteam.holiday.WebServer.Model;
 using myteam.holiday.WebServer.Services;
+using System.Security.Cryptography;
+using System.Text;
 
 namespace myteam.holiday.WebApi.Controllers
 {
@@ -9,14 +13,17 @@ namespace myteam.holiday.WebApi.Controllers
     public class UserController : ControllerBase
     {
         private readonly AppDbUserService _userService;
+        private readonly DbValidationService _validationService;
         private readonly ILogger<UserController> _logger;
 
         public UserController(
-            ILogger<UserController> logger,
-            AppDbUserService userService)
+            AppDbUserService userService,
+            DbValidationService validationService,
+            ILogger<UserController> logger)
         {
             _logger = logger;
             _userService = userService;
+            _validationService = validationService;
         }
 
         [HttpGet("GetAllUser")]
@@ -26,28 +33,57 @@ namespace myteam.holiday.WebApi.Controllers
         }
 
         [HttpGet("GetUser")]
-        public async Task<ActionResult<User>> GetUser(int userId)
+        public async Task<ActionResult<User>> GetUser(string login, string hash)
         {
-            return Ok(await _userService.GetOneAsync(userId));
+            if(await _validationService.IsConvergeUserPasswords(login, hash))
+            {
+                User user = await _userService.GetOneLoginAsync(login) ?? new();
+                user.Salt = default;
+                return Ok(user);
+            }
+            return BadRequest("Ошибка: некорректные данные модели User");
         }
 
         [HttpPost("CreateUser")]
-        public async Task<ActionResult<int>> CreateUser(User user)
+        public async Task<ActionResult<string>> CreateUser(User user, string password)
         {
-            return Ok(await _userService.CreateUserAsync(user));
+            if(await _validationService.IsValidCreateUserModelAsync(user))
+            {
+                byte[] salt = RandomNumberGenerator.GetBytes(128 / 8);
+                string hashed = Convert.ToBase64String(KeyDerivation.Pbkdf2(
+                    password: password,
+                    salt: salt,
+                    prf: KeyDerivationPrf.HMACSHA256,
+                    iterationCount: 100000,
+                    numBytesRequested: 256 / 8));
+                user.Hash = hashed;
+                user.Salt = Encoding.UTF8.GetString(salt);
+                user.RoleId = 1;
+                await _userService.CreateUserAsync(user);
+                return Ok(hashed);
+            }
+            return BadRequest("Ошибка: некорректные данные модели User");
         }
 
         [HttpPost("UpdateUser")]
-        public async Task<ActionResult<int>> UpdateUser(int oldId, User newUser)
+        public async Task<ActionResult<int>> UpdateUser(User newUser, string hash)
         {
-            return Ok(await _userService.UpdateUserAsync(oldId, newUser));
+            if(await _validationService.IsValidUpdateUserModelAsync(newUser, hash))
+            {
+                return Ok(await _userService.UpdateUserAsync(newUser));
+            }
+            return BadRequest("Ошибка: некорректные данные модели User");
         }
 
         [HttpDelete("DeleteUser")]
-        public async Task<ActionResult<int>> DeleteUser(int id)
+        public async Task<ActionResult<int>> DeleteUser(string login, string hash)
         {
-            User user = await _userService.GetOneAsync(id) ?? new();
-            return Ok(await _userService.DeleteUserAsync(user));
+            if (await _validationService.IsConvergeUserPasswords(login, hash))
+            {
+                User user = await _userService.GetOneLoginAsync(login) ?? new();
+                return Ok(await _userService.DeleteUserAsync(user));
+            }
+            return BadRequest("Ошибка: некорректные данные модели User");
         }
     }
 }
